portfolio/src/components/server/requirements.txt
Flask
python-dotenv
sendgrid